optimize code: add routes and used store
