+++ Dockerfile
@@ -1,13 +1,14 @@
-FROM node:14-alpine AS development
-ENV NODE_ENV development
-# Add a work directory
+FROM node:16-alpine
+
+LABEL maintainer="Leonardo Caycho"
+
 WORKDIR /app
-# Cache and Install dependencies
-COPY package.json .
-RUN npm install
-# Copy app files
-COPY . .
-# Expose port
+#tiempo del servidor igual al del sistema
+ENV TZ=UTC
+#para que los archivos puedan actualizarse en tiempo real al modificarlos en el entorno de docker
+ENV CHOKIDAR_USEPOLLING=true
+
+#colocando el tiempo del servidor
+RUN ln -snf /usr/share/zoneinfo/$TZ /etc/localtime && echo $TZ > /etc/timezone
+
 EXPOSE 3000
-# Start the app
-CMD [ "npm", "start" ]
