+++ docker-compose.dev.yml
@@ -1,13 +1,11 @@
-version: "3.8"
-
+version: '3'
 services:
-  app:
-    container_name: meet-your-congress-app-dev
-    image: meet-your-congress-app-dev
-    build:
-      context: .
-      target: development
-    volumes:
-      - ./src:/app/src
+  node:
+    build: .
     ports:
-      - 3000:3000
+      - '3000:3000'
+    volumes:
+      - '.:/app'
+    entrypoint: "sh -c"
+    command:
+      - 'npm install; npm start;'
