+++ src/components/AppContent.js
@@ -1,12 +1,15 @@
-import React, {Fragment, useState} from "react";
+import React, {Fragment, useState, Suspense} from "react";
 import {
     CButton,
     CCol,
     CContainer, CModal, CModalBody, CModalFooter, CModalHeader, CModalTitle,
-    CRow
+    CRow, CSpinner
 } from "@coreui/react";
 import classes from "./AppContent.module.css";
-import MemberCard from "./base/MemberCard";
+import MemberCard from "./base/utilities/MemberCard";
+import {Redirect, Route, Switch} from "react-router-dom";
+
+import routes from '../routes';
 
 /**
  * Shape formation of congress by #Diego
@@ -14,134 +17,34 @@ import MemberCard from "./base/MemberCard";
  * @return MainPage - First page of this project.
  * **/
 const AppContent = (props) => {
-    const classesHeader = `sidebar-fixed context ${classes["main-content"]}`;
-    const [visible, setVisible] = useState(false);
 
-    const LiveDemo = () => {
-
-    }
-
-    const congressMen = props.congressList;
-
-    const onMouseEnterInfoMan = (event, getIdParty) => {
-        if (!event.target.id.match("^[a-zA-Z]*$") && !event.target.id.match("^[0-9]*$")) {
-
-            props.showInfoParty(event.target.id, getIdParty);
-
-        }
-    };
-
-    const onClickInfoMan = (getIdMan) => {
-        setVisible(true);
-    };
-
-    let rowMax = 16;
-    let colMax = 15;
-    let row = 0;
-    let col = 0;
-    let spaceCounter = 6;
-    let patternCounter = 1;
-    let use = 1;
-
-    let middleCounter = 0;
-    let iterationCounter = 0;
-    let countFullSpace = 0;
-
-    let countCongress = 0;
-    let middle = ((colMax - 1) / 2);
-
-    const emptyCol = (id) => {
-        return <CCol key={id}/>;
-    };
-    const listCol = (row) => {
-
-        let congressRow = [];
-
-        if (row % 2 === 0 && row > 0) {
-            patternCounter = patternCounter + 2;
-            spaceCounter = spaceCounter - 1;
-            iterationCounter = iterationCounter + 1;
-            if (patternCounter === middle + 4) {
-                spaceCounter = spaceCounter - 1;
-            }
-        }
-
-        if (row % iterationCounter === 0) {
-            if (patternCounter > middle) {
-                if (patternCounter !== middle + 2) {
-                    middleCounter = middleCounter + 1 + use;
-
-                    if (use !== 0)
-                        use = use - 1;
-                }
-
-
-            }
-        }
-
-        for (col = 0; col < colMax; col++) {
-            if (col < spaceCounter || col >= colMax - spaceCounter) {
-                congressRow.push(emptyCol(countFullSpace));
-            }
-
-            if (col >= spaceCounter && col < colMax - spaceCounter) {
-                if (middle > col - middleCounter && middle < col + middleCounter) {
-                    congressRow.push(emptyCol(countFullSpace));
-                } else {
-                    const getMan = congressMen[countCongress];
-
-                    const getIdParty = getMan !== undefined ? getMan.idParty : null;
-                    const getIdMan = getMan !== undefined ? getMan.id : countCongress;
-                    const getColorMan = getMan !== undefined ? getMan.color : 'gray';
-                    const getImageMan = getMan !== undefined ? getMan.image : '/avatars/1.jpg';
-
-                    const values = (
-                        <Fragment>
-                            <CCol key={getIdMan}>
-                                    <MemberCard id={getIdMan}  key={getIdMan} idParty={getIdParty}
-                                                colorParty={getColorMan} srcImage={getImageMan}
-                                    onChooseMember={onMouseEnterInfoMan}/>
-                            </CCol>
-                        </Fragment>);
-                    countCongress = countCongress + 1;
-                    congressRow.push(values);
-                }
-            }
-            countFullSpace = countFullSpace + 1;
-        }
-
-        return congressRow;
-
-    };
-
-    const adds = [];
-    let congress = "";
-
-    const rows = () => {
-        for (row = 0; row < rowMax; row++) {
-            congress =
-                (<CRow key={row} lg={{cols: 15, gutter: 15}}>
-                    {listCol(row)}
-                </CRow>);
-
-            adds.push(congress);
-        }
-        return adds;
-
-    };
 // fluid='true'
     return (
         <Fragment>
-            <div className={classesHeader}>
-                <CContainer>
-                    <CRow lg={{cols: 1, gutter: 1}}>
-                        <h1 className='text-lg-center'>Congreso de la República</h1>
-                    </CRow>
-                    {rows()}
-                </CContainer>
-            </div>
+            <Suspense fallback={<CSpinner color="primary" />}>
+                <Switch>
+                    {routes.map((route, idx) => {
+                        return (
+                            route.component && (
+                                <Route
+                                    key={idx}
+                                    path={route.path}
+                                    exact={route.exact}
+                                    name={route.name}
+                                    render={(props) => (
+                                        <>
+                                            <route.component {...props} />
+                                        </>
+                                    )}
+                                />
+                            )
+                        )
+                    })}
+                    <Redirect from="/" to="/congress"/>
+                </Switch>
+            </Suspense>
         </Fragment>
     );
 };
 
-export default AppContent;
+export default React.memo(AppContent);
