+++ src/components/AppHeader.js
@@ -7,6 +7,7 @@ import {
     CNavLink,
 } from "@coreui/react";
  import {useDispatch, useSelector} from "react-redux";
+import {NavLink} from "react-router-dom";
 
 
 const AppHeader = () => {
@@ -14,15 +15,14 @@ const AppHeader = () => {
     const classesHeader = `bg-light ${classes.header}`;
     const [visible, setVisible] = useState(false);
 
-  //  const sidebarShow = useSelector((state) => state.sidebarShow);
+    const sidebarShow = useSelector((state) =>  state.sidebarShow);
 //    const dispatch = useDispatch({ type: 'set', sidebarShow: !sidebarShow });
     return (
         <CHeader position="sticky" className="mb-4">
             <CContainer fluid>
                 <CHeaderToggler
                     className="ms-md-3 d-lg-none"
-                    onClick={() => false} //dispatch({ type: 'set', sidebarShow: !sidebarShow })}
-                >
+                    onClick={() => dispatch({ type: 'set', sidebarShow: !sidebarShow })}>
                     <CIcon name="cil-menu" size="lg" />
                 </CHeaderToggler>
                 <CHeaderBrand className="mx-auto d-md-none" to="/">
@@ -30,13 +30,16 @@ const AppHeader = () => {
                 </CHeaderBrand>
                 <CHeaderNav className="d-none d-md-flex me-auto">
                     <CNavItem>
-                        <CNavLink href="#">Agenda Documentada</CNavLink>
+                        <CNavLink to="/congress" component={NavLink} activeClassName="active">Congreso</CNavLink>
                     </CNavItem>
                     <CNavItem>
-                        <CNavLink href="#">Decretos de Urgencia</CNavLink>
+                        <CNavLink to="/documented-agenda" component={NavLink}>Agenda Documentada</CNavLink>
                     </CNavItem>
                     <CNavItem>
-                        <CNavLink href="#">Proyectos de Ley</CNavLink>
+                        <CNavLink to="/emergency-decrees" component={NavLink}>Decretos de Urgencia</CNavLink>
+                    </CNavItem>
+                    <CNavItem>
+                        <CNavLink to="/bills-law" component={NavLink}>Proyectos de Ley</CNavLink>
                     </CNavItem>
                 </CHeaderNav>
             </CContainer>
