+++ src/components/base/Agenda.js
@@ -1,0 +1,8 @@
+
+
+
+const Agenda = () => {
+
+};
+
+export default Agenda;
