+++ src/components/base/BillsLaw.js
@@ -1,0 +1,7 @@
+
+
+const BillsLaw = () => {
+
+};
+
+export default BillsLaw;
