+++ src/components/base/EmergencyDecrees.js
@@ -1,0 +1,8 @@
+const EmergencyDecrees = () => {
+
+};
+
+export default EmergencyDecrees;
+
+
+
