+++ src/store.js
@@ -2,6 +2,7 @@ import { createStore } from 'redux';
 
 const initialState = {
   sidebarShow: false,
+  infoPartyShow: false,
 }
 
 const changeState = (state = initialState, { type, ...rest }) => {
@@ -12,6 +13,7 @@ const changeState = (state = initialState, { type, ...rest }) => {
       return state
   }
 };
+
 
 const store = createStore(changeState);
 export default store;
